ml-pretraining-app/requirements.txt
numpy<2.0.0
scipy<2.0.0
pandas>=1.3.5,<2.0.0
scikit-learn>=1.0.2,<2.0.0
PyYAML==6.0
matplotlib>=3.5.1,<4.0.0
pdfplumber==0.10.2
python-magic==0.4.27
transformers==4.36.2
torch==2.1.2
datasets==2.16.1
peft==0.7.1
accelerate==0.25.0
bitsandbytes==0.41.3